[update, add] 一旦コミット（コンパイルエラーあり）
- ステータスシステムとバフシステムを分離
- ステータスのベース値とバフ計算後の値を別々のオブジェクトで管理するように
- まだ未完成
