--- Assets/Scripts/InGame/Common/Stats/EffectableStatus.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d150c5b8eb945464db0144def6d2277b
