+++ Assets/Scripts/InGame/Common/Stats/PlayerParameter.cs
@@ -1,3 +1,4 @@
+using September.InGame.Common.Stats;
 using UnityEngine;
 
 namespace InGame.Player
@@ -18,5 +19,20 @@ namespace InGame.Player
         public float StaminaRegen => _staminaRegen;
         public float StaminaConsumption => _staminaConsumption;
         public int AttackDamage => _attackDamage;
+
+        public StatsContainer GetStats()
+        {
+            var stats = new Stat[]
+            {
+                new(StatType.Health, Health, maxValue: Health),
+                new(StatType.Speed, Speed),
+                new(StatType.Stamina, Stamina, maxValue: Stamina),
+                new(StatType.StaminaRegen, StaminaRegen),
+                new(StatType.StaminaConsumption, StaminaConsumption),
+                new(StatType.AttackDamage, AttackDamage),
+            };
+            
+            return new StatsContainer(stats);
+        }
     }
 }
