+++ Assets/Scripts/InGame/Common/Stats/StatsManager.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7430647939924cbda893a2234a9552d9
+timeCreated: 1777114652
