+++ Assets/Scripts/InGame/Common/Stats/StatsModifierBase.cs
@@ -1,0 +1,10 @@
+namespace September.InGame.Common.Stats
+{
+    /// <summary>
+    /// ステータスを変化させるクラス
+    /// </summary>
+    public abstract class StatsModifierBase
+    {
+        public abstract StatsContainer Apply(StatsContainer stats);
+    }
+}
