+++ Assets/Scripts/InGame/Common/Stats/StatsModifierBase.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 81ebef44df074db694e72d9679d6fda2
+timeCreated: 1777114849
